+++ make-release.ps1
@@ -2,7 +2,7 @@
 # Aufruf: .\make-release.ps1 [-Version "0.4"]
 
 param(
-    [string]$Version = "1.1.3"
+    [string]$Version = "1.1.4"
 )
 
 $ErrorActionPreference = 'Stop'